+++ src/main/java/org/mmarini/wheelly/model/ProxySample.java
@@ -31,13 +31,29 @@ package org.mmarini.wheelly.model;
 
 import io.reactivex.rxjava3.schedulers.Timed;
 
+import java.awt.geom.Point2D;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
+
+import static java.lang.Double.parseDouble;
+import static java.lang.Integer.parseInt;
+import static java.lang.Math.*;
+import static org.mmarini.wheelly.model.ScannerMap.normalizeAngle;
 
 /**
  *
  */
 public class ProxySample {
     private static final int NO_PARAMS = 7;
+
+    /**
+     * @param relativeDirection relative direction DEG
+     * @param distance          distance (m)
+     * @param asset             asset
+     */
+    public static ProxySample create(int relativeDirection, double distance, RobotAsset asset) {
+        return new ProxySample(relativeDirection, distance, asset);
+    }
 
     /**
      * The string status is formatted as:
@@ -54,17 +70,17 @@ public class ProxySample {
             throw new IllegalArgumentException("Missing status parameters");
         }
         long instant = clock.fromRemote(Long.parseLong(params[1]));
-        int relDirection = Integer.parseInt(params[2]);
-        float distance = Float.parseFloat(params[3]);
-        float x = Float.parseFloat(params[4]);
-        float y = Float.parseFloat(params[5]);
-        int direction = Integer.parseInt(params[6]);
+        int relDirection = parseInt(params[2]);
+        double distance = parseDouble(params[3]);
+        double x = parseDouble(params[4]);
+        double y = parseDouble(params[5]);
+        int direction = parseInt(params[6]);
         return new Timed<>(new ProxySample(relDirection, distance, RobotAsset.create(x, y, direction)),
                 instant, TimeUnit.MILLISECONDS);
     }
 
     public final RobotAsset asset;
-    public final float distance;
+    public final double distance;
     public final int relativeDirection;
 
     /**
@@ -72,9 +88,31 @@ public class ProxySample {
      * @param distance          distance (m)
      * @param asset             asset
      */
-    protected ProxySample(int relativeDirection, float distance, RobotAsset asset) {
+    protected ProxySample(int relativeDirection, double distance, RobotAsset asset) {
         this.relativeDirection = relativeDirection;
         this.distance = distance;
         this.asset = asset;
+    }
+
+    /**
+     * Returns the absolute location of the reflector if any
+     */
+    public Optional<Point2D> getLocation() {
+        if (distance > 0) {
+            Point2D location = asset.getLocation();
+            double angle = asset.getRadDirection() + toRadians(relativeDirection);
+            double x = location.getX() + distance * cos(angle);
+            double y = location.getY() + distance * sin(angle);
+            return Optional.of(new Point2D.Double(x, y));
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    /**
+     * Returns the sample direction in RAD
+     */
+    public double getSampleDirection() {
+        return normalizeAngle(asset.getRadDirection() + toRadians(relativeDirection));
     }
 }